This is an weather.com website scraper made with python
this file works with a decent internet connection it scrapes all the information from the website(www.weather.com) it will scrape all the information about temperature of all states in india,and it will show the wind speed in which direction ,humidity one main thing is please don't use this for long time. any modifications can be accepted.